fix llava rlhf
